refactor: rename services and add types
