+++ src/interfaces/interface.ts
@@ -4,3 +4,10 @@ export interface IDataCard {
   height: string;
   mass: string;
 }
+
+export interface IData {
+  count: number;
+  next: string | null;
+  previous: string | null;
+  results: [key: string];
+}
